requirements.txt
python==3.9.16
pytorch==1.12.1
numpy==1.23.5
scipy==1.10.0
scikit-learn==1.2.1
click==8.0.4
ruamel.yaml==0.17.21
logzero==1.7.0